@@ -6,7 +6,7 @@ readme = "README.md"
 authors = [
     { name = "Erwin Kuhn", email = "erwin.kuhn@protonmail.com" }
 ]
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 dependencies = ["pulse-framework>=0.1.18"]
 
 [project.scripts]
