@@ -3,7 +3,7 @@ name = "pulse-workspace"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 
 
 [dependency-groups]
@@ -51,6 +51,9 @@ members = [
 [tool.ruff.format]
 indent-style = "tab"
 
+[tool.ruff]
+target-version = "py312"
+
 [tool.ruff.lint]
 select = ["E4", "E7", "E9", "F", "B", "I"] # default + "I"
 
@@ -62,4 +65,5 @@ markers = [
 # Skip CDK tests by default (they're slow due to CDK synthesis)
 # Run them explicitly with: pytest -m cdk
 addopts = "-m 'not cdk'"
+norecursedirs = ["vendors", "node_modules", ".git"]
 asyncio_default_fixture_loop_scope = "function"
