@@ -46,7 +46,7 @@ export function serialize(data: Serializable): Serialized {
 
 		if (value instanceof Date) {
 			dates.push(idx);
-			return value.getTime();
+			return value.toISOString();
 		}
 
 		if (Array.isArray(value)) {
@@ -123,7 +123,20 @@ export function deserialize<Data extends Serializable = Serializable>(
 		}
 
 		if (dates.has(idx)) {
-			const dt = new Date(value as number);
+			if (typeof value !== "string") {
+				throw new Error("Date payload must be an ISO string");
+			}
+			const literal = parseDateLiteral(value);
+			if (literal) {
+				const [y, m, d] = literal;
+				const dt = new Date(Date.UTC(y, m - 1, d));
+				objects.push(dt);
+				return dt;
+			}
+			if (isDateLiteral(value)) {
+				throw new Error(`Invalid date literal: ${value}`);
+			}
+			const dt = new Date(value);
 			objects.push(dt);
 			return dt;
 		}
@@ -185,4 +198,35 @@ export function deserialize<Data extends Serializable = Serializable>(
 	}
 
 	return reconstruct(data);
+}
+
+const DATE_LITERAL_RE = /^\d{4}-\d{2}-\d{2}$/;
+
+function isDateLiteral(value: string): boolean {
+	return DATE_LITERAL_RE.test(value);
+}
+
+function parseDateLiteral(value: string): [number, number, number] | null {
+	if (!isDateLiteral(value)) {
+		return null;
+	}
+	const [yStr, mStr, dStr] = value.split("-");
+	const y = Number(yStr);
+	const m = Number(mStr);
+	const d = Number(dStr);
+	if (!Number.isInteger(y) || !Number.isInteger(m) || !Number.isInteger(d)) {
+		return null;
+	}
+	if (m < 1 || m > 12 || d < 1 || d > 31) {
+		return null;
+	}
+	const dt = new Date(Date.UTC(y, m - 1, d));
+	if (
+		dt.getUTCFullYear() !== y ||
+		dt.getUTCMonth() !== m - 1 ||
+		dt.getUTCDate() !== d
+	) {
+		return null;
+	}
+	return [y, m, d];
 }
