@@ -9,21 +9,84 @@ from typing import (
 	ParamSpec,
 	TypeAlias,
 	TypeVar,
+	final,
+	override,
 )
 
-from pulse.state import State
+from pulse.state.state import State
 
 T = TypeVar("T")
 TState = TypeVar("TState", bound="State")
 P = ParamSpec("P")
 R = TypeVar("R")
 
-QueryKey: TypeAlias = tuple[Hashable, ...]
-"""Tuple of hashable values identifying a query in the store.
+
+@final
+class Key(tuple[Hashable, ...]):
+	"""Normalized query key with a precomputed hash."""
+
+	_hash: int = 0
+
+	def __new__(cls, key: "QueryKey"):
+		if isinstance(key, Key):
+			return key
+		if isinstance(key, (list, tuple)):
+			parts = tuple(key)
+			try:
+				key_hash = hash(parts)
+			except TypeError as e:
+				raise TypeError(
+					f"Query key contains unhashable value: {e}.\n\n"
+					+ "Keys must contain only hashable values (strings, numbers, tuples).\n"
+					+ f"Got: {key!r}\n\n"
+					+ "If using a dict or list inside the key, convert it to a tuple:\n"
+					+ "    key=('users', tuple(user_ids))  # instead of list"
+				) from None
+			obj = super().__new__(cls, parts)
+			obj._hash = key_hash
+			return obj
+		raise TypeError(
+			f"Query key must be a tuple or list, got {type(key).__name__}: {key!r}\n\n"
+			+ "Examples of valid keys:\n"
+			+ "    key=('users',)           # single-element tuple\n"
+			+ "    key=('user', user_id)    # tuple with dynamic value\n"
+			+ "    key=['posts', 'feed']    # list form also works"
+		)
+
+	@override
+	def __hash__(self) -> int:
+		return self._hash
+
+
+QueryKey: TypeAlias = tuple[Hashable, ...] | list[Hashable] | Key  # pyright: ignore[reportImplicitStringConcatenation]
+"""List/tuple of hashable values identifying a query in the store.
 
 Used to uniquely identify queries for caching, deduplication, and invalidation.
-Keys are hierarchical tuples like ``("user", user_id)`` or ``("posts", "feed")``.
+Keys are hierarchical lists/tuples like ``("user", user_id)`` or ``["posts", "feed"]``.
+Lists are normalized to a tuple-backed Key internally.
 """
+
+
+def normalize_key(key: QueryKey) -> Key:
+	"""Convert a query key to a normalized key for use as a dict key."""
+	return Key(key)
+
+
+@final
+@dataclass(frozen=True, slots=True)
+class QueryKeys:
+	"""Wrapper for selecting multiple query keys."""
+
+	keys: tuple[Key, ...]
+
+	def __init__(self, *keys: QueryKey):
+		object.__setattr__(self, "keys", tuple(normalize_key(key) for key in keys))
+
+
+def keys(*query_keys: QueryKey) -> QueryKeys:
+	"""Create a QueryKeys wrapper for filtering by multiple keys."""
+	return QueryKeys(*query_keys)
+
 
 QueryStatus: TypeAlias = Literal["loading", "success", "error"]
 """Current status of a query.
