@@ -27,17 +27,20 @@ from pulse.queries.common import (
 	ActionError,
 	ActionResult,
 	ActionSuccess,
+	Key,
 	OnErrorFn,
 	OnSuccessFn,
 	QueryKey,
 	QueryStatus,
 	bind_state,
+	normalize_key,
 )
 from pulse.queries.query import RETRY_DELAY_DEFAULT, QueryConfig
 from pulse.reactive import Computed, Effect, Signal, Untrack
 from pulse.reactive_extensions import ReactiveList, unwrap
 from pulse.scheduling import TimerHandleLike, create_task, later
-from pulse.state import InitializableProperty, State
+from pulse.state.property import InitializableProperty
+from pulse.state.state import State
 
 T = TypeVar("T")
 TParam = TypeVar("TParam")
@@ -115,6 +118,7 @@ class RefetchPage(Generic[T, TParam]):
 	fetch_fn: Callable[[TParam], Awaitable[T]]
 	param: TParam
 	observer: "InfiniteQueryResult[T, TParam] | None" = None
+	clear: bool = False
 	future: "asyncio.Future[ActionResult[T | None]]" = field(
 		default_factory=asyncio.Future
 	)
@@ -141,7 +145,7 @@ class InfiniteQueryConfig(QueryConfig[list[Page[T, TParam]]], Generic[T, TParam]
 class InfiniteQuery(Generic[T, TParam], Disposable):
 	"""Paginated query that stores data as a list of Page(data, param)."""
 
-	key: QueryKey
+	key: Key
 	cfg: InfiniteQueryConfig[T, TParam]
 
 	@property
@@ -248,7 +252,7 @@ class InfiniteQuery(Generic[T, TParam], Disposable):
 		gc_time: float = 300.0,
 		on_dispose: Callable[[Any], None] | None = None,
 	):
-		self.key = key
+		self.key = normalize_key(key)
 
 		self.cfg = InfiniteQueryConfig(
 			retries=retries,
@@ -305,7 +309,8 @@ class InfiniteQuery(Generic[T, TParam], Disposable):
 
 		for obs in self._observers:
 			if obs._on_success is not None:  # pyright: ignore[reportPrivateUsage]
-				await maybe_await(call_flexible(obs._on_success, self.pages))  # pyright: ignore[reportPrivateUsage]
+				with Untrack():
+					await maybe_await(call_flexible(obs._on_success, self.pages))  # pyright: ignore[reportPrivateUsage]
 
 	async def _commit_error(self, error: Exception):
 		"""Commit error state and run error callbacks."""
@@ -313,7 +318,8 @@ class InfiniteQuery(Generic[T, TParam], Disposable):
 
 		for obs in self._observers:
 			if obs._on_error is not None:  # pyright: ignore[reportPrivateUsage]
-				await maybe_await(call_flexible(obs._on_error, error))  # pyright: ignore[reportPrivateUsage]
+				with Untrack():
+					await maybe_await(call_flexible(obs._on_error, error))  # pyright: ignore[reportPrivateUsage]
 
 	def _commit_sync(self):
 		"""Synchronous commit - updates state based on current pages."""
@@ -427,7 +433,8 @@ class InfiniteQuery(Generic[T, TParam], Disposable):
 		self._cancel_observer_actions(observer)
 
 		if len(self._observers) == 0:
-			self.schedule_gc()
+			if not self.__disposed__:
+				self.schedule_gc()
 
 	def invalidate(
 		self,
@@ -703,8 +710,8 @@ class InfiniteQuery(Generic[T, TParam], Disposable):
 
 		page = await action.fetch_fn(action.param)
 
-		if idx is None:
-			# Page doesn't exist - jump to this page, clearing existing pages
+		if action.clear or idx is None:
+			# clear=True or page doesn't exist - replace all pages with just this one
 			self.pages.clear()
 			self.pages.append(Page(page, action.param))
 		else:
@@ -782,12 +789,13 @@ class InfiniteQuery(Generic[T, TParam], Disposable):
 		*,
 		observer: "InfiniteQueryResult[T, TParam] | None" = None,
 		cancel_fetch: bool = False,
+		clear: bool = False,
 	) -> ActionResult[T | None]:
 		"""
 		Refetch a page by its param. Queued for sequential execution.
 
-		If the page doesn't exist, clears existing pages and loads the requested
-		page as the new starting point.
+		If the page doesn't exist or clear=True, clears existing pages and loads
+		the requested page as the new starting point.
 
 		Note: Prefer calling refetch_page() on InfiniteQueryResult to ensure the
 		correct fetch function is used. When called directly on InfiniteQuery, uses
@@ -795,7 +803,7 @@ class InfiniteQuery(Generic[T, TParam], Disposable):
 		"""
 		fn = fetch_fn if fetch_fn is not None else self.fn
 		action: RefetchPage[T, TParam] = RefetchPage(
-			fetch_fn=fn, param=param, observer=observer
+			fetch_fn=fn, param=param, observer=observer, clear=clear
 		)
 		return await self._enqueue(action, cancel_fetch=cancel_fetch)
 
@@ -1019,12 +1027,22 @@ class InfiniteQueryResult(Generic[T, TParam], Disposable):
 		page_param: TParam,
 		*,
 		cancel_fetch: bool = False,
+		clear: bool = False,
 	) -> ActionResult[T | None]:
+		"""Fetch a specific page by its param.
+
+		Args:
+			page_param: The page parameter to fetch.
+			cancel_fetch: Cancel any in-flight fetches before starting.
+			clear: If True, clears all other pages and keeps only the fetched page.
+				Useful for resetting pagination to a specific page.
+		"""
 		return await self._query().refetch_page(
 			page_param,
 			fetch_fn=self._fetch_fn,
 			observer=self,
 			cancel_fetch=cancel_fetch,
+			clear=clear,
 		)
 
 	def set_initial_data(
@@ -1149,7 +1167,7 @@ class InfiniteQueryProperty(Generic[T, TParam, TState], InitializableProperty):
 		Callable[[TState, list[Page[T, TParam]]], TParam | None] | None
 	)
 	_max_pages: int
-	_key: QueryKey | Callable[[TState], QueryKey] | None
+	_key: Key | Callable[[TState], Key] | None
 	# Not using OnSuccessFn and OnErrorFn since unions of callables are not well
 	# supported in the type system. We just need to be careful to use
 	# call_flexible to invoke these functions.
@@ -1193,7 +1211,17 @@ class InfiniteQueryProperty(Generic[T, TParam, TState], InitializableProperty):
 		self._on_success_fn = None
 		self._on_error_fn = None
 		self._initial_data = MISSING
-		self._key = key
+		if key is None:
+			self._key = None
+		elif callable(key):
+			key_fn = key
+
+			def normalized_key(state: TState) -> Key:
+				return normalize_key(key_fn(state))
+
+			self._key = normalized_key
+		else:
+			self._key = normalize_key(key)
 		self._initial_data_updated_at = initial_data_updated_at
 		self._enabled = enabled
 		self._fetch_on_mount = fetch_on_mount
@@ -1204,7 +1232,11 @@ class InfiniteQueryProperty(Generic[T, TParam, TState], InitializableProperty):
 			raise RuntimeError(
 				f"Cannot use @{self.name}.key decorator when a key is already provided to @infinite_query(key=...)."
 			)
-		self._key = fn
+
+		def normalized_key(state: TState) -> Key:
+			return normalize_key(fn(state))
+
+		self._key = normalized_key
 		return fn
 
 	def on_success(self, fn: OnSuccessFn[TState, list[T]]):
@@ -1278,8 +1310,17 @@ class InfiniteQueryProperty(Generic[T, TParam, TState], InitializableProperty):
 		)
 
 		if self._key is None:
+			# pyright: ignore[reportImplicitStringConcatenation]
 			raise RuntimeError(
-				f"key is required for infinite query '{self.name}'. Provide a key via @infinite_query(key=...) or @{self.name}.key decorator."
+				f"Missing query key for @infinite_query '{self.name}'. "
+				+ "A key is required to cache and share query results.\n\n"
+				+ f"Fix: Add key=(...) to the decorator or use the @{self.name}.key decorator:\n\n"
+				+ "    @ps.infinite_query(initial_page_param=..., key=('my_query',))\n"
+				+ f"    async def {self.name}(self, param): ...\n\n"
+				+ "Or with a dynamic key:\n\n"
+				+ f"    @{self.name}.key\n"
+				+ f"    def _{self.name}_key(self):\n"
+				+ "        return ('my_query', self.some_param)"
 			)
 		raw_initial = (
 			call_flexible(self._initial_data, state)
@@ -1384,6 +1425,7 @@ class InfiniteQueryProperty(Generic[T, TParam, TState], InitializableProperty):
 def infinite_query(
 	fn: Callable[[TState, TParam], Awaitable[T]],
 	*,
+	key: QueryKey | Callable[[TState], QueryKey] | None = None,
 	initial_page_param: TParam,
 	max_pages: int = 0,
 	stale_time: float = 0.0,
@@ -1395,7 +1437,6 @@ def infinite_query(
 	initial_data_updated_at: float | dt.datetime | None = None,
 	enabled: bool = True,
 	fetch_on_mount: bool = True,
-	key: QueryKey | None = None,
 ) -> InfiniteQueryProperty[T, TParam, TState]: ...
 
 
@@ -1403,6 +1444,7 @@ def infinite_query(
 def infinite_query(
 	fn: None = None,
 	*,
+	key: QueryKey | Callable[[TState], QueryKey] | None = None,
 	initial_page_param: TParam,
 	max_pages: int = 0,
 	stale_time: float = 0.0,
@@ -1414,7 +1456,6 @@ def infinite_query(
 	initial_data_updated_at: float | dt.datetime | None = None,
 	enabled: bool = True,
 	fetch_on_mount: bool = True,
-	key: QueryKey | None = None,
 ) -> Callable[
 	[Callable[[TState, Any], Awaitable[T]]],
 	InfiniteQueryProperty[T, TParam, TState],
@@ -1424,6 +1465,7 @@ def infinite_query(
 def infinite_query(
 	fn: Callable[[TState, TParam], Awaitable[T]] | None = None,
 	*,
+	key: QueryKey | Callable[[TState], QueryKey] | None = None,
 	initial_page_param: TParam,
 	max_pages: int = 0,
 	stale_time: float = 0.0,
@@ -1435,7 +1477,6 @@ def infinite_query(
 	initial_data_updated_at: float | dt.datetime | None = None,
 	enabled: bool = True,
 	fetch_on_mount: bool = True,
-	key: QueryKey | None = None,
 ) -> (
 	InfiniteQueryProperty[T, TParam, TState]
 	| Callable[
@@ -1449,7 +1490,8 @@ def infinite_query(
 	pagination. Data is stored as a list of pages, each with its data and the
 	parameter used to fetch it.
 
-	Requires ``@query_prop.key`` and ``@query_prop.get_next_page_param`` decorators.
+	Requires a key (``key=`` or ``@query_prop.key``) and
+	``@query_prop.get_next_page_param`` decorator.
 
 	Args:
 		fn: The async method to decorate (when used without parentheses).
