@@ -35,7 +35,7 @@ from pulse.scheduling import (
 	TimerRegistry,
 	create_future,
 )
-from pulse.state import State
+from pulse.state.state import State
 from pulse.transpiler.id import next_id
 from pulse.transpiler.nodes import Expr
 
@@ -112,7 +112,7 @@ class RouteMount:
 	) -> None:
 		self.render = render
 		self.path = ensure_absolute_path(path)
-		self.route = RouteContext(route_info, route)
+		self.route = RouteContext(route_info, route, render)
 		self.effect = None
 		self._pulse_ctx = None
 		self.tree = RenderTree(route.render())
@@ -288,7 +288,6 @@ class RenderSession:
 		self._send_message = None
 		self._global_states = {}
 		self._global_queue = []
-		self.query_store = QueryStore()
 		self.connected = False
 		self.channels = ChannelsManager(self)
 		self.forms = FormRegistry(self)
@@ -298,6 +297,7 @@ class RenderSession:
 		self._ref_channels_by_route = {}
 		self._tasks = TaskRegistry(name=f"render:{id}")
 		self._timers = TimerRegistry(tasks=self._tasks, name=f"render:{id}")
+		self.query_store = QueryStore()
 		self.prerender_queue_timeout = prerender_queue_timeout
 		self.detach_queue_timeout = detach_queue_timeout
 		self.disconnect_queue_timeout = disconnect_queue_timeout
@@ -581,9 +581,10 @@ class RenderSession:
 	# ---- Helpers ----
 
 	def close(self):
+		# Close all pending timers at the start, to avoid anything firing while we clean up
+		self._timers.cancel_all()
 		self.forms.dispose()
 		self._tasks.cancel_all()
-		self._timers.cancel_all()
 		for path in list(self.route_mounts.keys()):
 			self.detach(path, timeout=0)
 		self.route_mounts.clear()
@@ -606,6 +607,8 @@ class RenderSession:
 		self._pending_js_results.clear()
 		self._ref_channel = None
 		self._ref_channels_by_route.clear()
+		# Close any timer that may have been scheduled during cleanup (ex: query GC)
+		self._timers.cancel_all()
 		self._global_queue = []
 		self._send_message = None
 		self.connected = False
