@@ -7,7 +7,9 @@ from pulse.hooks.setup import setup, setup_key
 from pulse.hooks.stable import stable
 from pulse.hooks.state import state
 from pulse.reactive import Signal
-from pulse.state import State
+from pulse.render_session import RenderSession
+from pulse.routing import Route, RouteContext, RouteInfo, RouteTree
+from pulse.state.state import State
 
 
 class DummyState(State):
@@ -24,6 +26,31 @@ class DummyState(State):
 	@property
 	def dispose_calls(self) -> int:
 		return self._dispose_calls
+
+
+def make_route_info(
+	pathname: str, *, query_params: dict[str, str] | None = None, hash: str = ""
+) -> RouteInfo:
+	return {
+		"pathname": pathname,
+		"hash": hash,
+		"query": "",
+		"queryParams": query_params or {},
+		"pathParams": {},
+		"catchall": [],
+	}
+
+
+def make_route_context(route_info: RouteInfo):
+	def render():
+		return ps.div()
+
+	route = Route("/", ps.component(render))
+	routes = RouteTree([route])
+	session = RenderSession("test", routes)
+	route_ctx = RouteContext(route_info, route, session)
+	app = ps.App(routes=[route])
+	return app, session, route_ctx, route
 
 
 def test_setup_returns_same_value_within_context():
@@ -354,3 +381,18 @@ def test_stable_returns_consistent_wrappers():
 	with ctx:
 		with pytest.raises(KeyError):
 			stable("missing")
+
+
+def test_route_returns_route_info():
+	app, session, route_ctx, _route = make_route_context(make_route_info("/"))
+	with ps.PulseContext(app=app, render=session, route=route_ctx):
+		info = ps.route()
+		assert info["pathname"] == "/"
+		assert info["queryParams"] == {}
+
+
+def test_pulse_route_returns_definition():
+	app, session, route_ctx, route = make_route_context(make_route_info("/"))
+	with ps.PulseContext(app=app, render=session, route=route_ctx):
+		definition = ps.pulse_route()
+		assert definition is route
