@@ -124,6 +124,15 @@ class TestLambda:
 		code = emit(fn)
 		assert code == "function get_const_1() {\nreturn () => 42;\n}"
 
+	def test_lambda_in_subscript_target(self):
+		@javascript
+		def set_item(arr):
+			arr[lambda x: x] = 1
+
+		fn = set_item.transpile()
+		code = emit(fn)
+		assert code == "function set_item_1(arr) {\narr[x => x] = 1;\n}"
+
 
 # =============================================================================
 # List Comprehensions
@@ -192,6 +201,17 @@ class TestComprehensions:
 			== "function double_values_1(pairs) {\nreturn new Map(pairs.map(([k, v]) => [k, v * 2]));\n}"
 		)
 
+	def test_list_comp_in_subscript_target(self):
+		@javascript
+		def set_item(obj, items: Iterable[int]):
+			obj[[x for x in items]] = 1
+
+		fn = set_item.transpile()
+		code = emit(fn)
+		assert (
+			code == "function set_item_1(obj, items) {\nobj[items.map(x => x)] = 1;\n}"
+		)
+
 
 # =============================================================================
 # Format Specs
@@ -211,7 +231,7 @@ class TestFormatSpecs:
 		code = emit(fn)
 		assert (
 			code
-			== "function format_pi_1() {\nlet pi = 3.14159;\nreturn `${pi.toFixed(2)}`;\n}"
+			== "function format_pi_1() {\nlet pi;\npi = 3.14159;\nreturn `${pi.toFixed(2)}`;\n}"
 		)
 
 	def test_left_align(self):
